+++ standingsrequests/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "standingsrequests.apps.StandingsRequestsConfig"
 
-__version__ = "0.5.0a7"
+__version__ = "0.5.0a8"
 __title__ = "Standings Requests"