+++ standingsrequests/admin.py
@@ -1,14 +1,12 @@
 from django.contrib import admin
 
 from .models import (
-    ContactSet,
     StandingRequest,
     StandingRevocation,
     EveEntity,
     CharacterContact,
     CorporationContact,
 )
-from .tasks import standings_update
 
 
 class AbstractStandingsRequestAdmin(admin.ModelAdmin):
@@ -60,34 +58,3 @@ class StandingsRequestAdmin(AbstractStandingsRequestAdmin):
 @admin.register(StandingRevocation)
 class StandingsRevocationAdmin(AbstractStandingsRequestAdmin):
     pass
-
-
-@admin.register(EveEntity)
-class EveEntityAdmin(admin.ModelAdmin):
-    list_display = ("entity_id", "name", "category", "updated")
-    list_filter = ("category",)
-
-    def has_add_permission(self, request):
-        return False
-
-    def has_change_permission(self, request, obj=None):
-        return False
-
-
-@admin.register(ContactSet)
-class ContactSetAdmin(admin.ModelAdmin):
-    list_display = ("date",)
-
-    actions = ("start_update",)
-
-    def start_update(self, request, queryset):
-        standings_update.delay()
-        self.message_user(request, "Fetching new contact set / standings...")
-
-    start_update.short_description = "Fetch new contact set / standings"
-
-    def has_add_permission(self, request):
-        return False
-
-    def has_change_permission(self, request, obj=None):
-        return False